+++ db/migrate/20160326112357_drop_currencies.rb
@@ -1,5 +1,0 @@
-class DropCurrencies < ActiveRecord::Migration
-  def change
-    drop_table :currencies
-  end
-end