+++ db/migrate/20160504154438_add_created_at_index_to_tweets.rb
@@ -1,5 +1,0 @@
-class AddCreatedAtIndexToTweets < ActiveRecord::Migration
-  def change
-    add_index :tweets, :created_at
-  end
-end