+++ db/migrate/20160504154515_add_created_at_index_to_articles.rb
@@ -1,5 +1,0 @@
-class AddCreatedAtIndexToArticles < ActiveRecord::Migration
-  def change
-    add_index :articles, :created_at
-  end
-end