+++ lib/mysql_client.rb
@@ -1,0 +1,28 @@
+require 'mysql2'
+
+def import_rates(rate_file)
+  client = Mysql2::Client.new(Settings.mysql)
+  query = File.read(File.join(Settings.application_root, 'aggregate/import.sql'))
+  client.query(query.gsub('$FILE', rate_file))
+  client.close
+end
+
+def get_rates(date)
+  client = Mysql2::Client.new(Settings.mysql)
+  query = File.read(File.join(Settings.application_root, 'aggregate/backup.sql'))
+  result = client.query(query.gsub('$DAY', date.strftime('%F')))
+  client.close
+  result
+end
+
+def create_candle_sticks(param)
+  begin
+    client = Mysql2::Client.new(Settings.mysql)
+    query = File.read(File.join(Settings.application_root, 'aggregate.sql'))
+    param.each {|key, value| query.gsub!("$#{key.upcase}", value) }
+    client.query(query)
+  rescue
+  ensure
+    client.close
+  end
+end