+++ script/rates/delete.sql
@@ -1,4 +1,0 @@
-DELETE FROM
-  rates
-WHERE
-  time < '$TIME'