+++ script/tweets/delete.sql
@@ -1,4 +1,0 @@
-DELETE FROM
-  tweets
-WHERE
-  created_at < '$CREATED_AT'